1)Fixed Bind tests 2) Added Provision
Repaired the test cases for Bind refactored them to use BindCommand
because it simplifies it a ton so I don't have to parse the url
manually.
Fixed bind because the id's were in opposite spots before and added
extra error handling so the user is required to pass in whatever the
BindCommand struct requires.